Scenegraph rewrite fixes plus fixed an out of bounds memory access (thanks valgrind): Fixed: * particles aren't drawn (no tire smoke) * input graph doesn't display * racing line doesn't have correct transparency * track map doesn't display

+++ src/k1999.cpp
@@ -229,7 +229,7 @@ void K1999::CalcRaceLine()
 	const unsigned int stepsize = 128;
 	
 	//abort if the track isn't long enough
-	if ((int)tx.size() < stepsize)
+	if (tx.size() < stepsize)
 		return;
 	
  //
